Fixing regex for identifying IMAP inbox folders
In the `update` method in the EmailBox class, there is a regex for checking the returned mail folders from the IMAP_SSL connection. However, this regex does not catch the case where in an inbox is specified in the format `(\\HasNoChildren) "." INBOX`. This is the format returned by my IMAP server and is valid.

This commit modifies the regex to allow either the '/' or '.' to separate the flags from the folder name.